// src/app/models/consulta.ts
export class Consulta {

    _id: number;
    dniPaciente: number;
    nombrePaciente: string;
    motivo: string;
    matricula: string;
    nombreMedico: string;
    especialidad: string;
    fecha: Date;
    link: string;
    estado: string;
    motivoRechazo: string;
    
    Consulta(){

    }
}